Change backend endpoint request body to match with frontend

+++ backend/src/main/java/de/phamkv/projectdiary/controller/PostController.java
@@ -3,11 +3,14 @@ package de.phamkv.projectdiary.controller;
 import de.phamkv.projectdiary.model.Post;
 import de.phamkv.projectdiary.model.Profile;
 import de.phamkv.projectdiary.service.PostService;
+import de.phamkv.projectdiary.service.ProfileService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/posts")
@@ -15,8 +18,11 @@ public class PostController {
 
     private final PostService postService;
 
-    public PostController(PostService postService) {
+    private final ProfileService profileService;
+
+    public PostController(PostService postService, ProfileService profileService) {
         this.postService = postService;
+        this.profileService = profileService;
     }
 
     @GetMapping("")
@@ -35,7 +41,14 @@ public class PostController {
     }
 
     @PostMapping("")
-    public ResponseEntity<Post> addPost(@RequestBody Post post) {
+    public ResponseEntity<Post> addPost(Authentication authentication, @RequestBody Post post) {
+        String username = authentication.getName();
+
+        Profile profile = profileService.getProfileByUsername(username);
+        System.out.println(username);
+
+        post.setProfile(profile);
+
         Post newPost = postService.addPost(post);
         return ResponseEntity.status(HttpStatus.CREATED).body(newPost);
     }
@@ -56,7 +69,15 @@ public class PostController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deletePost(@PathVariable Long id) {
+    public ResponseEntity<Void> deletePost(Authentication authentication, @PathVariable Long id) {
+        String username = authentication.getName();
+
+        Post post = postService.getPostById(id);
+
+        if (!username.equals(post.getProfile().getUsername())) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
         postService.deletePost(id);
         return ResponseEntity.noContent().build();
     }
